Remove unused files and update configurations for cleaner project structure
