--- client/_headers
@@ -1,12 +1,0 @@
-/*.js
-  Content-Type: text/javascript
-/*.mjs
-  Content-Type: text/javascript
-/*.ts
-  Content-Type: text/javascript
-/*.tsx
-  Content-Type: text/javascript
-/*.jsx
-  Content-Type: text/javascript
-/assets/*.js
-  Content-Type: text/javascript
