+++ client/src/main.tsx
@@ -1,4 +1,4 @@
-import { StrictMode } from 'react'
+
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.tsx'
@@ -7,10 +7,10 @@ import { ToastContainer } from 'react-toastify'
 
 
 createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <BrowserRouter basename='/Diagram-Datawarehouse'>
+
+    <BrowserRouter >
     <App />
     <ToastContainer />
     </BrowserRouter>
-  </StrictMode>,
+
 )
