+++ client/src/pages/UploadPage.tsx
@@ -5,8 +5,7 @@ import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 import { motion } from "framer-motion";
 
-// const VITE_BASE_URL = 'https://harshsaw.tech/datadiagram/api';
-const VITE_BASE_URL = 'http://localhost:4001/api';
+const VITE_BASE_URL = import.meta.env.VITE_BASE_URL || 'http://localhost:4000/api';
 
 // Updated interfaces to match the new API response structure
 interface DiagramType {
