--- test/requirements.txt
@@ -1,5 +1,0 @@
-selenium
-webdriver-manager
-requests
-webdriver-manager 
-pillow
